Add files via upload
Description :
Given the details on customers’ orders and promotion emails sent, the task is to predict whether the customer is churned or retained. Also, we study the effect of email promotion on the customer churn rate, so that we can explain whether upsell or cross-selling of products improves the customer relationship

Steps followed in the project:
Data understanding, Data cleaning and preprocessing, EDA (univariate, bivariate, mutli variate), Statistical Test, Class imbalance treatment, Feature Engineering, Base model, Modelling after SMOTE, Modelling after Feature Selection, Final Model using Hyperparameters tuning, Feature Imporatnce, conclusion

Suggestions:
In order to retain the customers, the retailers adapted various marketing strategies and market basket analysis and various other methods to study their customer shopping pattern. As per our literature survey, we have seen that the customer churn was predicted mainly based on CRM records, feedbacks from social media, feedback survey conducted on behalf of the company. But our perspective is that most of the marketing is done through primary mode of communication which is email. So, in this project, we are going to analyze whether the promotional emails sent to the customer has any effect on customer churn. Based on the analysis, we suggest that, offers and freebies can be offered for customer who will be churned. More customised offers can be given to them. Further if we get the data on products that they have purchased, we can perform market basket analysis and provide them better recommendations by Identifying the patterns of shopping in the customers